Decrease def batch size to 800 for SQLite compat.
On some install/version there seems to be a limit of 999 to the
number of parameters of SQLite statements. Since we rely on a
big "id IN (...)" list of ID for batching, it's safer to use 800
as the default batch size. Anyway one is free to increase this
limit if needed.